# docs/contracts/chat-console.openapi.yaml
openapi: 3.1.0
info:
  title: Sprint Crew console API
  version: 0.1.0
  description: |
    Status: Implemented. Console sessions are durable in SQLite (M1) and reaped after
    CONSOLE_SESSION_TTL_DAYS. Clarify questions are model-generated by the Interpreter
    on the Work lane (ADR 0013) and degrade to a deterministic fallback when the lane
    is cold.
    Contract for the interactive console
    (docs/contracts/chat-console-api.md, ADR 0012, ADR 0013). Routes are live at
    /v1/console/* alongside the existing /sprint/* and /health API.
    Auth: when CONSOLE_API_TOKEN is set, every /v1/console/* and /sprint/* request
    requires Authorization: Bearer <token>, or ?token= as a fallback on any route
    (needed for EventSource, which cannot set headers). Empty token disables auth.
    GET /health is exempt.
    Timeline: GET /events polls a durable, monotonically sequenced log (M2) and
    GET /stream serves the same events over SSE with Last-Event-ID resume (M3).
    Runs are queued one at a time and cancellable (M5, ADR 0014).
# Either scheme satisfies any route: api/auth.py accepts ?token= everywhere, not just on
# the SSE stream, so a generated client should treat both as valid globally.
security:
  - bearerAuth: []
  - tokenQueryAuth: []
paths:
  /v1/console/sessions:
    post:
      operationId: createConsoleSession
      summary: Create a console session (with initial_prompt → clarifying or ready)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateConsoleSessionRequest"
            example:
              mode: code
              initial_prompt: Add a /metrics endpoint with request counters
              repo_url: https://github.com/example/service
              target_language: null
      responses:
        "201":
          description: |
            Session created. With initial_prompt the Interpreter (or fallback) runs
            immediately — status is clarifying or ready, not collecting.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ConsoleSession"
              example:
                session_id: cs-7f3a
                mode: code
                status: clarifying
                confirmed: false
                repo_url: https://github.com/example/service
                target_language: null
                messages:
                  - role: user
                    content: Add a /metrics endpoint with request counters
                    timestamp: "2026-07-16T09:00:00+00:00"
                intent: null
                clarify_questions:
                  - question_id: q-scope
                    text: Which part of the repo should change?
                    why_asked: null
                    recommended_suggestion_id: null
                    suggestions:
                      - suggestion_id: s-scope-focused
                        label: Only the files needed for this change
                      - suggestion_id: s-scope-broad
                        label: Related modules too
                        detail: including tests and docs touched by the change
                    allow_custom: true
                clarify_answers: []
                sprint_ref: null
                plan_result: null
                error: null
                workspace_status: cloning
                workspace_root: null
                workspace_error: null
                index_status: pending
                index_error: null
                created_at: "2026-07-16T09:00:00+00:00"
                updated_at: "2026-07-16T09:00:00+00:00"
        "422":
          $ref: "#/components/responses/ValidationError"
    get:
      operationId: listConsoleSessions
      summary: Session history, newest first (M8)
      description: |
        Summaries only — no messages, clarify state or plan results. Fetch one session
        by id for the full object.
      parameters:
        - name: limit
          in: query
          schema: { type: integer, default: 25, minimum: 1, maximum: 100 }
        - name: offset
          in: query
          schema: { type: integer, default: 0, minimum: 0 }
      responses:
        "200":
          description: One page of session summaries.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ConsoleSessionPage"
              example:
                sessions:
                  - session_id: cs-7f3a
                    mode: code
                    status: completed
                    workspace_status: ready
                    title: Add a /metrics endpoint with request counters
                    repo_url: https://github.com/example/service
                    created_at: "2026-07-16T09:00:00+00:00"
                    updated_at: "2026-07-16T09:41:00+00:00"
                total: 1
                next_offset: null
  /v1/console/sessions/{id}:
    get:
      operationId: getConsoleSession
      summary: Fetch session state (polling endpoint)
      parameters:
        - $ref: "#/components/parameters/SessionId"
      responses:
        "200":
          description: Current session state
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ConsoleSession"
        "404":
          $ref: "#/components/responses/NotFound"
    delete:
      operationId: deleteConsoleSession
      summary: Delete a session and everything it owns (M8)
      description: |
        Removes the session row, its clone, its diff snapshots and its timeline.
        Refuses while a run is live — cancel first; deleting under a running agent
        would leave it writing into a workspace nobody owns.
      parameters:
        - $ref: "#/components/parameters/SessionId"
      responses:
        "204":
          description: Deleted.
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          description: Session is queued, running or awaiting_review.
  /v1/console/sessions/{id}/messages:
    post:
      operationId: postConsoleMessage
      summary: Append a user chat message and re-interpret
      description: |
        No longer inert outside `collecting` (M9). A message posted while clarifying or
        ready re-runs the Interpreter over the whole conversation, which REPLACES the open
        question set: `clarify_round` increments, `clarify_questions` is a brand new set
        with new round-scoped ids, previously given answers move to `prior_clarifications`,
        and `confirmed` is reset because the confirmation belonged to a different
        understanding of the request.

        Consequences for a client state machine: `ready` can go BACK to `clarifying` — the
        first backwards transition here — `clarify_questions` is not append-only, and an
        answered question can stop being answered. Re-render the clarify form from the
        response rather than merging into what was on screen.

        409 while an ask is in flight: the Interpreter and the Explainer want the same lane.
      parameters:
        - $ref: "#/components/parameters/SessionId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PostMessageRequest"
            example:
              content: It should also cover the backlog endpoints
      responses:
        "200":
          description: Updated session (may transition to clarifying, in either direction)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ConsoleSession"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/StateConflict"
        "422":
          $ref: "#/components/responses/ValidationError"
  /v1/console/sessions/{id}/clarify:
    post:
      operationId: submitClarifyAnswers
      summary: Answer pending clarify questions
      description: |
        Question ids are round-scoped since M9 (`q-{round}-{n}`). Answering an id from a
        superseded round is 409, not 400: the request was correct when it was rendered, and
        the fix is to refetch the session and answer the set that is open now.
      parameters:
        - $ref: "#/components/parameters/SessionId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ClarifyRequest"
            example:
              answers:
                - question_id: q-1-scope
                  selected_suggestion_id: s-1-scope-focused
                  custom_text: null
                - question_id: q-1-tests
                  selected_suggestion_id: null
                  custom_text: unit tests only, no live tiers
      responses:
        "200":
          description: Updated session (ready once all questions answered)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ConsoleSession"
        "400":
          description: Unknown question_id, both/neither answer fields, or custom answer where allow_custom is false
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ConsoleError"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          description: |
            Session is not clarifying, or the answered question belongs to a superseded
            clarify round — refetch and answer the current one.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ConsoleError"
        "422":
          $ref: "#/components/responses/ValidationError"
  /v1/console/sessions/{id}/confirm:
    post:
      operationId: confirmConsoleSession
      summary: Explicitly confirm the clarified request (requires status=ready)
      parameters:
        - $ref: "#/components/parameters/SessionId"
      responses:
        "200":
          description: Session with confirmed=true (status stays ready)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ConsoleSession"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/StateConflict"
  /v1/console/sessions/{id}/start:
    post:
      operationId: startConsoleRun
      summary: Queue the run (requires status=ready and confirmed=true)
      description: |
        Returns 202 in milliseconds: the run is accepted, not performed. Workspace
        clone, vector indexing, context enrichment, the Work-lane load and the
        ScrumMaster call all happen afterward and report on the event stream.

        mode=code: drives today's from-prompt pipeline; session gains sprint_ref
        (backlog_run_id for GET /sprint/backlog/{run_id}; its session_ids feed
        GET /sprint/session/{session_id}). Status is "queued" with queue_position
        when another run holds the single run slot, otherwise "running".

        mode=plan: analysis only, never ships (ADR 0012). BREAKING in M10 — this
        used to complete inside the request and return plan_result. It is now a real
        ScrumMaster (and, at depth=deep, TechLead) run taking the same run slot as a
        code run, so it returns "queued"/"running" with plan_result still null. Watch
        for the plan_complete event and re-read the session, or read plan_run_id.
      parameters:
        - $ref: "#/components/parameters/SessionId"
      requestBody:
        required: false
        description: |
          Optional; the route took no body before M10, so omitting it stays valid.
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StartRunRequest"
      responses:
        "202":
          description: Run queued or running (both modes)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ConsoleSession"
              examples:
                codeModeStarted:
                  summary: mode=code — backlog run dispatched
                  value:
                    session_id: cs-7f3a
                    mode: code
                    status: running
                    confirmed: true
                    repo_url: https://github.com/example/service
                    target_language: null
                    messages: []
                    intent: null
                    clarify_questions: []
                    clarify_answers:
                      - question_id: q-scope
                        selected_suggestion_id: s-scope-focused
                        custom_text: null
                    sprint_ref:
                      backlog_run_id: run-91c2
                      sprint_session_ids: []
                    plan_result: null
                    error: null
                    queue_position: null
                    cancel_requested_at: null
                    created_at: "2026-07-16T09:00:00+00:00"
                    updated_at: "2026-07-16T09:05:00+00:00"
                codeModeQueued:
                  summary: mode=code — another run holds the single run slot
                  value:
                    session_id: cs-9b12
                    mode: code
                    status: queued
                    confirmed: true
                    repo_url: https://github.com/example/service
                    target_language: null
                    messages: []
                    intent: null
                    clarify_questions: []
                    clarify_answers: []
                    sprint_ref:
                      backlog_run_id: run-4de7
                      sprint_session_ids: []
                    plan_result: null
                    error: null
                    queue_position: 1
                    cancel_requested_at: null
                    created_at: "2026-07-16T09:06:00+00:00"
                    updated_at: "2026-07-16T09:06:00+00:00"
                planModeQueued:
                  summary: mode=plan — accepted; the plan arrives on the stream
                  value:
                    session_id: cs-plan1
                    mode: plan
                    status: running
                    confirmed: true
                    repo_url: null
                    target_language: null
                    messages:
                      - role: user
                        content: Split metrics work into backlog stories
                        timestamp: "2026-07-16T09:00:00+00:00"
                    intent: null
                    clarify_questions: []
                    clarify_answers: []
                    sprint_ref: null
                    plan_result: null
                    plan_run_id: plan-3c8e
                    parent_session_id: null
                    error: null
                    queue_position: null
                    cancel_requested_at: null
                    created_at: "2026-07-16T09:00:00+00:00"
                    updated_at: "2026-07-16T09:05:00+00:00"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          description: Not ready or not confirmed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ConsoleError"
              example:
                detail: session must be confirmed before start
...
  /v1/console/sessions/{id}/cancel:
    post:
      operationId: cancelConsoleSession
      summary: Cancel from any non-terminal state
      description: |
        Two outcomes, distinguished by the body rather than the status code.

        Nothing started, or the run is still queued: status becomes "cancelled"
        immediately. The run is executing: Stop is accepted — cancel_requested_at
        is set, a cancel_requested event goes out, and status stays "running"
        until the run unwinds at its next checkpoint. The Stop button therefore
        needs a pending state.

        A running cancel cannot be instant: a checkpoint is invisible while a
        subprocess is mid-call, so the worst case is bounded by
        ACCEPTANCE_TEST_TIMEOUT_S and run_command's own timeout.
      parameters:
        - $ref: "#/components/parameters/SessionId"
      responses:
        "200":
          description: Cancelled, or cancel accepted and pending
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ConsoleSession"
              examples:
                cancelledImmediately:
                  summary: Nothing running — terminal at once
                  value:
                    status: cancelled
                    cancel_requested_at: null
                    queue_position: null
                cancelPending:
                  summary: Run executing — stopping at the next checkpoint
                  value:
                    status: running
                    cancel_requested_at: "2026-07-16T09:10:00+00:00"
                    queue_position: null
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/StateConflict"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: opaque
      description: |
        Shared CONSOLE_API_TOKEN. Omit the Authorization header when the env is
        empty (auth disabled). GET /health never requires this scheme.
    tokenQueryAuth:
      type: apiKey
      in: query
      name: token
      description: |
        The same shared CONSOLE_API_TOKEN as bearerAuth, passed as a ?token= query
        parameter. Accepted on every route but intended only for GET .../stream, where
        browser EventSource cannot set an Authorization header. A query-string token
        may appear in access logs — acceptable for this single-user LAN deployment.
  parameters:
    SessionId:
      name: id
      in: path
      required: true
      schema:
        type: string
        minLength: 1
...
  responses:
    NotFound:
      description: Unknown session
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ConsoleError"
          example:
            detail: Console session not found
    StateConflict:
      description: Action not allowed in the session's current status
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ConsoleError"
    ValidationError:
      description: Request body failed schema validation (extra=forbid)
      content:
        application/json:
          schema:
            type: object
  schemas:
    ConsoleMode:
      type: string
      enum: [plan, code]
...
    CreateConsoleSessionRequest:
      type: object
      additionalProperties: false
      required: [mode]
      properties:
        mode:
          $ref: "#/components/schemas/ConsoleMode"
        initial_prompt:
          type: [string, "null"]
        repo_url:
          type: [string, "null"]
        target_language:
          type: [string, "null"]
          description: Phase 3 language-specialized lanes; pass null for now.
    StartRunRequest:
      type: object
      additionalProperties: false
      description: |
        Optional body for POST /start (M10). depth applies to plan mode only; a code
        run's depth is the pipeline's business, not the caller's. Omit for the
        server default (CONSOLE_PLAN_DEFAULT_DEPTH, "quick" out of the box).
      properties:
        depth:
          oneOf:
            - $ref: "#/components/schemas/PlanDepth"
            - type: "null"
    PostMessageRequest:
      type: object
      additionalProperties: false
      required: [content]
      properties:
        content:
          type: string
          minLength: 1
    ClarifyRequest:
      type: object
      additionalProperties: false
      required: [answers]
      properties:
        answers:
          type: array
          minItems: 1
          items:
            $ref: "#/components/schemas/ClarifyAnswer"
...
    ConsoleSessionPage:
      type: object
      additionalProperties: false
      required: [sessions, total]
      properties:
        sessions:
          type: array
          items:
            $ref: "#/components/schemas/ConsoleSessionSummary"
        total:
          type: integer
        next_offset:
          type: [integer, "null"]
          description: Offset of the next page, or null when this is the last one.
...
    ConsoleError:
      type: object
      additionalProperties: false
      required: [detail]
      properties:
        detail:
          type: string
          minLength: 1